code compiles and tests pass in Java 11

+++ super-csv/src/test/java/org/supercsv/cellprocessor/ParseDoubleTest.java
@@ -21,7 +21,7 @@ import static org.supercsv.SuperCsvTestUtils.ANONYMOUS_CSVCONTEXT;
 import org.junit.Before;
 import org.junit.Test;
 import org.supercsv.cellprocessor.ift.CellProcessor;
-import org.supercsv.exception.SuperCsvCellProcessorException;
+import org.supercsv.SuperCsvCellProcessorException;
 import org.supercsv.mock.IdentityTransform;
 
 /**
@@ -55,12 +55,12 @@ public class ParseDoubleTest {
 	@Test
 	public void testValidDoubles() {
 		// positive values
-		assertEquals(POSITIVE_VAL, processor.execute(POSITIVE_VAL, ANONYMOUS_CSVCONTEXT), 0.0);
-		assertEquals(POSITIVE_VAL, processorChain.execute(POSITIVE_VAL, ANONYMOUS_CSVCONTEXT), 0.0);
+		assertEquals(POSITIVE_VAL, (Double) processor.execute(POSITIVE_VAL, ANONYMOUS_CSVCONTEXT), 0.0);
+		assertEquals(POSITIVE_VAL, (Double) processorChain.execute(POSITIVE_VAL, ANONYMOUS_CSVCONTEXT), 0.0);
 		
 		// negative values
-		assertEquals(NEGATIVE_VAL, processor.execute(NEGATIVE_VAL, ANONYMOUS_CSVCONTEXT), 0.0);
-		assertEquals(NEGATIVE_VAL, processorChain.execute(NEGATIVE_VAL, ANONYMOUS_CSVCONTEXT), 0.0);
+		assertEquals(NEGATIVE_VAL, (Double) processor.execute(NEGATIVE_VAL, ANONYMOUS_CSVCONTEXT), 0.0);
+		assertEquals(NEGATIVE_VAL, (Double) processorChain.execute(NEGATIVE_VAL, ANONYMOUS_CSVCONTEXT), 0.0);
 	}
 	
 	/**
@@ -69,12 +69,12 @@ public class ParseDoubleTest {
 	@Test
 	public void testValidDoubleStrings() {
 		// positive values
-		assertEquals(POSITIVE_VAL, processor.execute(POSITIVE_STRING, ANONYMOUS_CSVCONTEXT), 0.0);
-		assertEquals(POSITIVE_VAL, processorChain.execute(POSITIVE_STRING, ANONYMOUS_CSVCONTEXT), 0.0);
+		assertEquals(POSITIVE_VAL, (Double) processor.execute(POSITIVE_STRING, ANONYMOUS_CSVCONTEXT), 0.0);
+		assertEquals(POSITIVE_VAL, (Double) processorChain.execute(POSITIVE_STRING, ANONYMOUS_CSVCONTEXT), 0.0);
 		
 		// negative values
-		assertEquals(NEGATIVE_VAL, processor.execute(NEGATIVE_STRING, ANONYMOUS_CSVCONTEXT), 0.0);
-		assertEquals(NEGATIVE_VAL, processorChain.execute(NEGATIVE_STRING, ANONYMOUS_CSVCONTEXT), 0.0);
+		assertEquals(NEGATIVE_VAL, (Double) processor.execute(NEGATIVE_STRING, ANONYMOUS_CSVCONTEXT), 0.0);
+		assertEquals(NEGATIVE_VAL, (Double) processorChain.execute(NEGATIVE_STRING, ANONYMOUS_CSVCONTEXT), 0.0);
 	}
 	
 	/**
